Working with Deferred Blocks
:+1:  we want to do is to add a new page that uses a deferred block – one of the coolest additions of Angular 17!
:100:  A deferred block is a block of code that is only rendered when a condition is met.
:a:  That means, we can use it to show a loading indicator or skeleton while we are loading data, and then show the actual data when it’s available.
Ok, that’s nothing new. We could do that before, right?
:+1:  Well, the great thing is that now the deferred template is not loaded until required!
This means, the items will not appear in the DOM before, which can massively improve performance.
There’s even a block @loading which describes the transition from empty state to the actual data, which makes it great for SSR.
:abc:  It’s mostly the same, but we now use the @defer directive to only render the list and infinite scroll when we are not loading anymore.
Additionally, we use the @placeholder directive to show the skeleton items while we are loading.
:accept:  The @error directive is used to show the error message, and the @loading directive is used to show a transition message while we go from placeholder to deferred block.
:airplane:  we learned how to use the new Angular Control Flow, Signals, and Deferred Blocks to build a modern Angular app with Ionic.
We also learned how to use the new Ionic standalone components to only include the components we actually use in our app.